requirements.txt
appdirs==1.4.3
apturl==0.5.2
autopep8==1.5.7
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
coloredlogs==7.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cwltool==2.0.20200224214940
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dominate==2.6.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
gpg===1.13.1-unknown
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
humanfriendly==4.18
idna==2.8
importlib-metadata==1.5.0
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
mypy-extensions==0.4.3
netifaces==0.10.4
networkx==2.4
newsapi-python==0.2.6
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
packaging==21.2
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pip-check-reqs==2.3.2
protobuf==3.6.1
prov==1.5.2
psutil==5.5.1
psycopg2==2.9.1
pycairo==1.16.2
pycodestyle==2.8.0
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyperclip==1.8.2
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
rdflib==4.2.2
rdflib-jsonld==0.4.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ruamel.yaml==0.15.89
schema-salad==5.0.20200220195218
SecretStorage==2.3.1
shellescape==3.4.1
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
SPARQLWrapper==1.8.5
SQLAlchemy==1.4.26
systemd-python==234
toml==0.10.2
typing-extensions==3.7.4.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
visitor==0.1.3
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.0.2
WTForms==2.3.3
xkit==0.0.0
zipp==1.0.0
